game.py : Core Rock-Paper-Scissors game
Player is given the option to pick one of the options of rock, paper and scissors.

The computer randomly picks one of the options of rock, paper and scissors.

Winner for each round is determined and one point is given to the winner.
The first to get five points wins the game. Once the game is finished, the player
is asked to quit or restart the game.

The current number of round and score card is displayed in every round.

Player can also quit the game at any time.