game.py: Code refactoring
The main game code is refactored. The main functionality of game is not
changed. Added the class so that unit testing can be done using unittest.

test.py: Unit testing module for the game

Testing code for various functionality if the game such as verifying the user
input, verifying the winner of the round, verifying the score management system
and verifying the game winner.